chore: restructure test project to match package
Added test/src subdirectories to test project to match directory
structure of package project. Moved source files to test/src
subdirectories. Updated namespaces to match file locations and package
namespaces under test. Updated project to reflect new file locations and
assembly name.
