+++ Chickensoft.GodotTestDriver.Tests/Chickensoft.GodotTestDriver.Tests.csproj
@@ -3,7 +3,7 @@
     <TargetFramework>net8.0</TargetFramework>
     <EnableDynamicLoading>true</EnableDynamicLoading>
     <LangVersion>preview</LangVersion>
-    <RootNamespace>GodotTestDriver.Tests</RootNamespace>
+    <RootNamespace>Chickensoft.GodotTestDriver.Tests</RootNamespace>
     <!-- Required for some nuget packages to work -->
     <!-- godotengine/godot/issues/42271#issuecomment-751423827 -->
     <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
