+++ Chickensoft.GodotTestDriver.Tests/project.godot
@@ -10,14 +10,14 @@ config_version=5
 
 [application]
 
-config/name="GodotTestDriver.Tests"
-run/main_scene="res://Tests.tscn"
+config/name="Chickensoft.GodotTestDriver.Tests"
+run/main_scene="res://test/Tests.tscn"
 config/features=PackedStringArray("4.4", "C#", "Mobile")
 config/icon="res://icon.svg"
 
 [dotnet]
 
-project/assembly_name="GodotTestDriver.Tests"
+project/assembly_name="Chickensoft.GodotTestDriver.Tests"
 
 [editor]
