+++ Chickensoft.GodotTestDriver.Tests/test/src/Drivers/GraphEditDriverTest.cs
@@ -1,4 +1,4 @@
-namespace Chickensoft.GodotTestDriver.Tests;
+namespace Chickensoft.GodotTestDriver.Tests.Drivers;
 
 using System.Linq;
 using Chickensoft.GoDotTest;
