+++ Chickensoft.GodotTestDriver.Tests/test/src/Drivers/Sprite2DDriverTest.tscn
